NPE solved, working on utlilities

+++ Home_works/Lection_8/Bookkeeping_demo/src/main/java/Interfaces/Manager.java
@@ -3,7 +3,7 @@ package Interfaces;
 import java.util.List;
 
 public interface Manager {
-    public int getManagerBonus();
+    public long getManagerBonus();
     public List<Employee> getManagersEmployees();
     public boolean removeEmployeeFromManager(Employee employeeForFiring);
     public boolean addEmployeeToManager(Employee employeeForHiring);
